+++ libclamav/readdb.c
@@ -1504,10 +1504,10 @@ static int cli_loadcbc(FILE *fs, struct cl_engine *engine, unsigned int *signo, 
 			struct cli_bc *bc0 = &bcs->all_bcs[i];
 			cli_errmsg("Can only load 1 BC_STARTUP bytecode, attempted to load 2nd!\n");
 			cli_warnmsg("Previous BC_STARTUP: %d %d by %s\n",
-				    bc0->id, bc0->metadata.timestamp,
+				    bc0->id, (uint32_t)bc0->metadata.timestamp,
 				    bc0->metadata.sigmaker ? bc0->metadata.sigmaker : "N/A");
 			cli_warnmsg("Conflicting BC_STARTUP: %d %d by %s\n",
-				    bc->id, bc->metadata.timestamp,
+				    bc->id, (uint32_t)bc->metadata.timestamp,
 				    bc->metadata.sigmaker ? bc->metadata.sigmaker : "N/A");
 			return CL_EMALFDB;
 		    }
